refactoring computation results and their commands
